+++ axum/src/routing/tests/handle_error.rs
@@ -29,7 +29,7 @@ impl<R> Service<R> for Svc {
     }
 }
 
-#[tokio::test]
+#[crate::test]
 async fn handler() {
     let app = Router::new().route(
         "/",
@@ -48,7 +48,7 @@ async fn handler() {
     assert_eq!(res.status(), StatusCode::REQUEST_TIMEOUT);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn handler_multiple_methods_first() {
     let app = Router::new().route(
         "/",
@@ -68,7 +68,7 @@ async fn handler_multiple_methods_first() {
     assert_eq!(res.status(), StatusCode::REQUEST_TIMEOUT);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn handler_multiple_methods_middle() {
     let app = Router::new().route(
         "/",
@@ -91,7 +91,7 @@ async fn handler_multiple_methods_middle() {
     assert_eq!(res.status(), StatusCode::REQUEST_TIMEOUT);
 }
 
-#[tokio::test]
+#[crate::test]
 async fn handler_multiple_methods_last() {
     let app = Router::new().route(
         "/",